Mon, Feb 16, 2026, 10:20 AM +01:00

+++ einstein/einstein.py
@@ -1,5 +1,5 @@
 def einstein(mass):
-    return float(mass)*300000000**2
+    return int(mass)*300000000**2
 
 mass = input()
 print(einstein(mass))
